requirements.txt
streamlit
tensorflow
Pillow
kaggle
scikit-learn
shutil